Moves schematests -> schematests/references
