+++ javatests/arcs/schematests/references/schema.arcs
@@ -1,5 +1,5 @@
 meta
-  namespace: arcs.schematests
+  namespace: arcs.schematests.references
 
 schema Level0
   name: Text
@@ -16,25 +16,25 @@ schema Level3
   name: Text
   children: [&Level2]
 
-particle Writer0 in 'arcs.schematests.Writer0'
+particle Writer0 in 'arcs.schematests.references.Writer0'
   level0: writes [Level0]
 
-particle Reader0 in 'arcs.schematests.Reader0'
+particle Reader0 in 'arcs.schematests.references.Reader0'
   level0: reads [Level0]
 
-particle Writer1 in 'arcs.schematests.Writer1'
+particle Writer1 in 'arcs.schematests.references.Writer1'
   level0: reads writes Level0
   level1: writes [Level1]
 
-particle Reader1 in 'arcs.schematests.Reader1'
+particle Reader1 in 'arcs.schematests.references.Reader1'
   level1: reads [Level1]
 
-particle Writer2 in 'arcs.schematests.Writer2'
+particle Writer2 in 'arcs.schematests.references.Writer2'
   level0: reads writes Level0
   level1: reads writes Level1
   level2: writes [Level2]
 
-particle Reader2 in 'arcs.schematests.Reader2'
+particle Reader2 in 'arcs.schematests.references.Reader2'
   level2: reads [Level2]
 
 @trigger
